Move all packages under Depends to Imports and fix resultant bugs
I've been calling MIMOSA inside another function, and I think that the
fact it loads the MASS, plyr, reshape, Biobase and ggplot2 libraries
(as they are under Depends in DESCRIPTION) was causing an issue with
another function. So, I moved these libraries to Imports, and my issue was
resolved.
In my understanding, it's better practice for a function not to cause
side effects (when they aren't the point of the function, anyway). So,
this change may be useful in general, and I think should probably be incorporated
into the code base.
There was one issue that arose from moving plyr to Imports, as the
ConstructExpressionSet function's .variable argument expected output
from a call to plyr::`.`, and now that isn't attached, breaking old code.
I added a fix that I think works, allowing old code that just calls `.` implicitly
to still work. Details below.
The vignette now needed ggplot2 manually loaded, and so that was done.

The tests all pass and the vignette produces the same output as before.

Changes per file:
- DESCRIPTION
  - Moved MASS, plyr, reshape, Biobase and ggplot2 from Depends to Imports
- NAMESPACE
  - Imported plyr::.
- R/MIMOSA.R
  - The only real difficulty I encountered with moving the packages from Depends
    to Imports was that for the `ConstructExpressionSet` function the intended argument
    to the `.variables` parameter used the `plyr::.` function and now plyr is no longer attached,
    which means that previous analysis scripts assuming no need to load plyr or explicitly specify
    the plyr library for `.` would break.
  - So, I added an `if` statement at the start of `ConstructExpressionSet` that would detect if
    the argument to `.variables` was a call to the `.` function without specifying the `plyr`
    library. If it was, then the call to `.` was replaced by a call to `plyr::.`. If not,
    then it was evaluated as normal. So either `.` or `plyr::`.` should work. So any previous code
    that functioned should still function.
- tests/*
  - I wrapped the call to MIMOSA in `test-MIMOSA.R` inside `suppressWarnings`. I get the
    following warning otherwise:
    "Formula does not contain the RefTreat variable. It will be added automatically. Set RT=FALSE to disable this."
  - There is just one whitespace difference for helper-init.R.
- vignettes/MIMOSA.Rnw
  - Since ggplot2 is no longer under Depends, it needed to be loaded manually at the start. Code added for this.